# AdidasSubscriptionService/swagger.yaml
swagger: '2.0'
info:
  description: Adidas Subscription API Description
  version: '1.0'
  title: Adidas Subscription API
  termsOfService: 'urn:tos'
  contact:
    name: Sergio Martinez Rodriguez
    url: 'https://github.com/sergiomarnez/adidasCodeChallenge/'
    email: sergiomarnez@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
...
consumes:
  - application/xml
  - application/json
produces:
  - application/xml
  - application/json
paths:
...
  /subscription/:
    post:
      tags:
        - subscription-service-controller
      summary: Creates a subscription on the system
      description: 'Saves subscription in database, sends an email to the subscriber confirming all it''s ok and create an event on the system. Also returns subscription ID created'
      operationId: createSubscriptionUsingPOST
      consumes:
        - application/xml
        - application/json
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: subs
          description: subs
          required: true
          schema:
            $ref: '#/definitions/Subscription'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /subscription/hey:
    get:
      tags:
        - subscription-service-controller
      summary: Say hey and display creator email
      operationId: heyUsingGET
      consumes:
        - application/xml
        - application/json
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  ObjectId:
    type: object
    properties:
      counter:
        type: integer
        format: int32
      date:
        type: string
        format: date-time
      machineIdentifier:
        type: integer
        format: int32
      processIdentifier:
        type: integer
        format: int32
      time:
        type: integer
        format: int64
      timeSecond:
        type: integer
        format: int32
      timestamp:
        type: integer
        format: int32
...
  Subscription:
    type: object
    properties:
      consent:
        type: boolean
      dateOfBirth:
        type: string
        format: date-time
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      id:
        $ref: '#/definitions/ObjectId'
      newsletterId:
        type: integer
        format: int32
...